Share Parquet page skipping tests with Delta Lake
